+++ wagtail/admin/templatetags/wagtailadmin_tags.py
@@ -31,9 +31,8 @@ from wagtail.admin.navigation import get_explorable_root_page
 from wagtail.admin.search import admin_search_areas
 from wagtail.admin.staticfiles import versioned_static as versioned_static_func
 from wagtail.admin.ui import sidebar
-from wagtail.admin.utils import get_admin_base_url
+from wagtail.admin.utils import get_admin_base_url, get_valid_next_url_from_request
 from wagtail.admin.views.bulk_action.registry import bulk_action_registry
-from wagtail.admin.views.pages.utils import get_valid_next_url_from_request
 from wagtail.admin.widgets import ButtonWithDropdown, PageListingButton
 from wagtail.coreutils import camelcase_to_underscore
 from wagtail.coreutils import cautious_slugify as _cautious_slugify
