+++ wagtail/admin/views/bulk_action/base_bulk_action.py
@@ -7,7 +7,7 @@ from django.views.generic import FormView
 
 from wagtail import hooks
 from wagtail.admin import messages
-from wagtail.admin.views.pages.utils import get_valid_next_url_from_request
+from wagtail.admin.utils import get_valid_next_url_from_request
 
 
 class BulkAction(ABC, FormView):
