+++ wagtail/admin/views/pages/utils.py
@@ -1,10 +1,2 @@
-from django.utils.http import url_has_allowed_host_and_scheme
-
-
-def get_valid_next_url_from_request(request):
-    next_url = request.POST.get("next") or request.GET.get("next")
-    if not next_url or not url_has_allowed_host_and_scheme(
-        url=next_url, allowed_hosts={request.get_host()}
-    ):
-        return ""
-    return next_url
+# Retain backwards compatibility for imports
+from wagtail.admin.utils import get_valid_next_url_from_request  # noqa
